# contracts/openapi/v1.locked.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Health Data Hub",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/apple/batch": {
      "post": {
        "description": "Receive a metric batch from HealthSave iOS app.\n\nExpected payload:\n{\n    \"metric\": \"heart_rate\",\n    \"batch_index\": 0,\n    \"total_batches\": 1,\n    \"samples\": [\n        {\"date\": \"2024-01-15T10:30:00Z\", \"qty\": 72, \"source\": \"Apple Watch\"},\n        ...\n    ]\n}",
        "operationId": "apple_batch_api_apple_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Batch"
      }
    },
    "/api/apple/status": {
      "get": {
        "description": "Return record counts so the iOS app knows what's synced.",
        "operationId": "apple_status_api_apple_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Status"
      }
    },
...
    "/api/v2/sync/coverage": {
      "get": {
        "description": "Return metric-level receipt coverage from the Data Hub side.",
        "operationId": "sync_coverage_api_v2_sync_coverage_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Coverage Api V2 Sync Coverage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Coverage"
      }
    },
    "/api/v2/sync/runs/latest": {
      "get": {
        "description": "Summarize the most recently observed HealthSave sync run.",
        "operationId": "latest_sync_run_api_v2_sync_runs_latest_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Latest Sync Run Api V2 Sync Runs Latest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Sync Run"
      }
    },
...
  }
}